% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{add_unique_id}
\alias{add_unique_id}
\title{Add unique ID for each voter in a race.}
\usage{
add_unique_id(df, state = "SC", year = "2010")
}
\arguments{
\item{df}{A dataset with county and voter ID}

\item{state}{state}
}
\description{
This implies adding county and voter_id within county
}
