% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{add_wi}
\alias{add_wi}
\title{Add "other" and write in votes}
\usage{
add_wi(tbl, var)
}
\arguments{
\item{tbl}{wide table}

\item{var}{the unquoted office variable}
}
\description{
For a given office, looks at office_vote (the candidate in characters) and
the office_party (the party, -1, 1, or otherwise 0), and edits the 0 to 0.5
if office_party is 0 but candiddate name is not NA (a third party or write-in)
}
