% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballotDef.R
\name{ballotDef}
\alias{ballotDef}
\title{ballotDef}
\usage{
ballotDef(name, ballot, combo)
}
\arguments{
\item{ballot}{a 'matrix' containing integer ballot results.}

\item{borders}{a 'matrix' defining border information. Either contiguity::csvTable or contiguity::borderTable format may be passed. csvTable input is converted to borderTable format.}
}
\description{
ballotDef
}
