% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballotDef.R
\name{ballotSortDef}
\alias{ballotSortDef}
\title{ballotSortDef}
\usage{
ballotSortDef(ballotDef, SBSORTFUN = sbChisqTest, n = 1000,
  decreasing = TRUE, sortChunkSize = 10000, MCLAPPLYFUN = get.lapply())
}
\arguments{
\item{SBSORTFUN}{a 'function' which takes a sub-ballot matrix and returns a sortable value. Default is sbChisqTest.}

\item{n}{an 'integer' the size of the head() to extract from the sorted SBSORTFUN results. Default is 1e3.}

\item{decreasing}{a 'logical' the order in which to sort SBSORTFUN results. Default is TRUE.}

\item{sortChunkSize}{an 'integer' defining the maximum number of combinations to process. Default is 1e4.}

\item{MCLAPPLYFUN}{a 'function' which performs lapply(). Default is the result of mclapplyFunGen() if it is installed, and lapply() otherwise.}
}
\description{
ballotSortDef
}
