% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballotTag.R
\name{ballotTag}
\alias{ballotTag}
\title{ballotTag}
\usage{
ballotTag(x, ballot = NULL, bTitle = NULL, bRegionTag = NULL)
}
\arguments{
\item{x}{a 'character' containing a label.}

\item{ballot}{a 'matrix' defining ballot information.}

\item{bTitle}{a 'character' containing a title}

\item{bRegionTag}{a 'regionTag'}
}
\description{
ballotTag
}
