% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{calculateSample}
\alias{calculateSample}
\alias{sbCalculateSample}
\title{sbCalculateSample}
\usage{
sbCalculateSample(sb, norm = FALSE)
}
\arguments{
\item{sb}{A sub-ballot}

\item{norm}{'logical' value. Default FALSE. Whether sample is to be normalised.}
}
\value{
a 'numeric' vector. The sample calculated from the sub-ballot
}
\description{
Calculate sample from sub-ballot
}
