% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{cand_to_ascii}
\alias{cand_to_ascii}
\title{Basic standardizations}
\usage{
cand_to_ascii(tbl)
}
\arguments{
\item{tbl}{A table with the column "cand_name"}
}
\description{
Removes a special character, change to ascii for standardization, remove periods at the end of names
}
