% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{cast_to_wide}
\alias{cast_to_wide}
\title{long voter-vote to voter-based data}
\usage{
cast_to_wide(df = raw, contests = c("PTY", "PRS", "USHOU", "USSEN",
  "GOV", "LGV", "SOS", "ATG", "SSI", "HOU", "SEN", "JPRB", "SHF", "COR",
  "CLR", "AUD", "CTRES", "CCL", "SCH", "JPRB", "WAT", "Q", "LRCA"))
}
\arguments{
\item{contests}{A vector of contest_codes to select and put into the wide.}

\item{raw}{A standardized long dataset with contest_code and contest_name}
}
\description{
This drops offices where multiple votes are cast for the same contest_code. Absentees are also dropped.
}
