% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{code_c_name}
\alias{code_c_name}
\title{Separate contest and number}
\usage{
code_c_name(tbl, ref_county, ref_state)
}
\arguments{
\item{tbl}{long data}

\item{ref_county}{dataframe with manual contest_codes for county-wide referenda to join}

\item{ref_state}{same, but for statewide referenda}
}
\description{
create a "code" that should be a seven-character code if properly formatted
reassemble contest by code and name
}
