% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{customSd}
\alias{customSd}
\title{Custom Standard Deviation}
\usage{
customSd(x, center = mean(x))
}
\arguments{
\item{x}{a 'numeric' vector. The sample}

\item{center}{a 'numeric' the center from which to calculate standard deviation. length(center)==1}
}
\value{
a 'numeric' The standard deviation of the sample, measured from the center.
}
\description{
Standard deviation function with a parameter to adjust center.
}
