% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{densityJobDef}
\alias{densityJobDef}
\title{densityJobDef}
\usage{
densityJobDef(ballot, combo, SBREPORTFUNS = list(sbDensityGen()),
  MCLAPPLYFUN = get.lapply(), .combine = list)
}
\description{
densityJobDef
}
