% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{fix_custom}
\alias{fix_custom}
\title{Custom fixes for contest names that are hard to recode by regex}
\usage{
fix_custom(tbl, elec)
}
\arguments{
\item{tbl}{long tibble  of votes}

\item{elec}{election year}
}
\description{
Custom fixes for contest names that are hard to recode by regex
}
