% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{G2016_Dillon_EL155}
\alias{G2016_Dillon_EL155}
\title{Sample EL155}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 106553 rows and 3 columns.}
\source{
South Carolina Election Commission
}
\usage{
data(G2016_Dillon_EL155)
}
\description{
Dillon County, SC EL155 file from the 2016 General Election read in by \env{read_EL155}.
}
\keyword{datasets}
