% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dataCombo.R
\name{get.dataCombo}
\alias{get.dataCombo}
\title{get.dataCombo}
\usage{
get.dataCombo(x, FUN = return, key = TRUE)
}
\arguments{
\item{x}{a 'ballotTag' object}

\item{FUN}{a 'function' or 'primitive' for the dataCombo. Default is
return()}

\item{key}{a 'vector' to key ballot columns. Default is TRUE}
}
\description{
get.dataCombo
}
