% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nb.R
\name{get.nb.layerTag}
\alias{get.nb.layerTag}
\title{get.nb.layerTag}
\usage{
\method{get.nb}{layerTag}(x)
}
\arguments{
\item{x}{a '<foo>Tag' object or 'character' which is coerced to layerTag}
}
\description{
get.nb.layerTag
}
