% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.simpleTest.R
\name{get.simpleTest.ballotTag}
\alias{get.simpleTest.ballotTag}
\title{get.simpleTest.ballotTag}
\usage{
\method{get.simpleTest}{ballotTag}(x, grepFilter = NULL, ...)
}
\description{
get.simpleTest.ballotTag
}
