% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Spatial.R
\name{get.Spatial.ballotTag}
\alias{get.Spatial.ballotTag}
\title{get.Spatial.ballotTag}
\usage{
\method{get.Spatial}{ballotTag}(x)
}
\arguments{
\item{x}{a '<foo>Tag' object or 'character' which is coerced to layerTag}
}
\description{
get.Spatial.ballotTag
}
