% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Spatial.R
\name{get.Spatial.layerTag}
\alias{get.Spatial.layerTag}
\title{get.Spatial.layerTag}
\usage{
\method{get.Spatial}{layerTag}(x)
}
\arguments{
\item{x}{a '<foo>Tag' object or 'character' which is coerced to layerTag}
}
\description{
get.Spatial.layerTag
}
