% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Spatial.R
\name{get.Spatial.regionTag}
\alias{get.Spatial.regionTag}
\title{get.Spatial.regionTag}
\usage{
\method{get.Spatial}{regionTag}(x)
}
\arguments{
\item{x}{a '<foo>Tag' object or 'character' which is coerced to layerTag}
}
\description{
get.Spatial.regionTag
}
