% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{get_precinct_range}
\alias{get_precinct_range}
\title{Extract precinct-identifying information}
\usage{
get_precinct_range(df)
}
\arguments{
\item{df}{a EL155 dataset, read through read_EL155}
}
\value{
A dataset keyed by precinct. start_id and end_id refer to the range of
the precinct in terms of the row IDs in read_EL155 output. column `p_name` is
is the name portion of the precinct header. `precinct_id` is the ID of
precinct within the county
}
\description{
Get a precinct-level dataset by using dplyr::distinct(.keep_all = TRUE).
Thus the rows need to be ordered correctly.
}
