% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_waffle.R
\name{gg_wfl}
\alias{gg_wfl}
\title{Waffle Plot of votes, colored}
\usage{
gg_wfl(
  tbl_indiv,
  var,
  nrows = 31,
  rev = FALSE,
  blank = FALSE,
  legend = FALSE,
  title_size = 0.8
)
}
\arguments{
\item{tbl_indiv}{A wide dataset of individual votes}

\item{var}{The variable in the dataset to use as the colors}

\item{nrows}{number of rows and columns}
}
\description{
Waffle Plot of votes, colored
}
\examples{
gg_wfl(filter(w_prez, PRS_party == 1), USH_party, nrows = 10)

}
