% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.nb.R
\name{group.nb}
\alias{group.nb}
\title{group.nb}
\usage{
group.nb(nb, subset)
}
\arguments{
\item{nb}{'nb' object}

\item{subset}{a 'vector' defining the currently occupied subset.}
}
\description{
Given the context of an 'nb' object and a
subset of occupied territory, return the territory outwith
the input subset that borders the input subset.
}
