% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{identify_voter}
\alias{identify_voter}
\title{assign individual voter ID by reading the asterisk mark}
\usage{
identify_voter(df)
}
\arguments{
\item{df}{a data frame where each row is a birth. The column `voter_top`
is 1 if the entry is the first vote on the ballot and 0 otherwise.}
}
\description{
Use cumsum to find jumps in the asterisk
}
