% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-.nb.R
\name{invade.nb}
\alias{invade.nb}
\title{invade.nb}
\usage{
invade.nb(nb, subset)
}
\arguments{
\item{nb}{'nb' object}

\item{subset}{a 'vector' defining the currently occupied subset.}
}
\description{
Given the context of an 'nb' object and a 
subset of occupied territory, return the territory outwith
the input subset that borders the input subset.
}
