% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-.nb.R
\name{invade.nb.test}
\alias{invade.nb.test}
\title{invade.nb.test}
\usage{
invade.nb.test(x, subset)
}
\arguments{
\item{subset}{a 'vector' defining the currently occupied subset.}
}
\description{
invade.nb.test
}
