% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is.valid.ballot}
\alias{is.valid.ballot}
\title{is.valid.ballot}
\usage{
is.valid.ballot(x, name = "")
}
\arguments{
\item{x}{an object to test}

\item{name}{a 'character' identifying the data under review}
}
\description{
Checks an object for validity as a matrix containing ballot information.
}
