% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_coding.R
\name{join_1col}
\alias{join_1col}
\title{Join with candidate data but single contest}
\usage{
join_1col(tbl, pattern, race = p_race, new_name, cand = NULL,
  na_thresh = 0.5, ...)
}
\arguments{
\item{tbl}{A wide dataset with the contest}

\item{pattern}{regex to mark that race}

\item{cand}{A key that matches candidate names to parties}

\item{...}{geographic variables to segment on}

\item{p_race}{A separate dataset that indicates the number of votes cast for
a contest code in a particular geography}
}
\description{
For example, used for referenda where there is only one "seat"
}
