% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_to_party.R
\name{join_dist}
\alias{join_dist}
\alias{join_county}
\alias{join_countydist}
\title{Join by a single district identifier (congerss, state hosue) or county}
\usage{
join_dist(tbl, cands, office)

join_county(tbl, cands, office)

join_countydist(tbl, cands, office)
}
\arguments{
\item{tbl}{variables must be of the type '{office}_{vote}', '{office}_{dist}' (if dist) and
county
for vote name and district, respectively.}

\item{cands}{A table of candidates. Must have variables \code{elec}, \code{dist},
\code{county}, \code{ballot_name}, \code{party_num}, \code{contest_type}}

\item{office}{NSE short form of the office in question}
}
\description{
Calls \link{join_1col} internally
}
\details{
Use \code{join_dist} for an office based on a single district number (like USHOU
or state house, use \code{join_county} for an office countywide (like Sheriff),
\code{join_countydist} for an office based on both (like county council). All are
wrappers to \code{join_1col}. We simplify the problem into just a one-variable
merge.

It creates new columns for each  \code{contest}: \code{contest_party}, a numeric
-1 (D), 0 (Abstain / 3rd Party / Write-in), 1 (R), and \code{contest_dist}, a
number for the district of the voter of the contest in question if applicable, and
\code{contest_ncand}, the number of D-Rs in the district (does not count 3rd party)
}
