% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_coding.R
\name{join_office}
\alias{join_office}
\title{Join and add number of candidates}
\usage{
join_office(long, cand, new_name, ...)
}
\arguments{
\item{long}{a long dataset}

\item{cand}{dataframe of candidates}

\item{new_name}{a symbol to use for the new dataset}
}
\description{
Join and add number of candidates
}
