% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerTag.R
\name{layerTag}
\alias{layerTag}
\title{layerTag}
\usage{
layerTag(x, lTitle = NULL, refInfo = NULL, dsn = NULL, layer = NULL,
  textLinks = NULL, linkSearchCol = NULL)
}
\arguments{
\item{x}{a 'layerTag'}

\item{lTitle}{a 'character' title}

\item{refInfo}{a 'character' roxygen2 @reference section}

\item{dsn}{a 'character' path to shapefile}

\item{layer}{a 'character' layer of shapefile}

\item{textLinks}{a 'character' vector defining links (ferries / bridges / etc)}

\item{linkSearchCol}{a reference to the field of the shapefile referenced by #'@export}
}
\description{
layerTag
}
