% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{normalise}
\alias{normalise}
\title{Sample normaliser}
\usage{
normalise(x, center = mean(x))
}
\arguments{
\item{x}{a 'numeric' vector. The sample to be normalised}

\item{center}{a 'numeric' the center to normalise to. length(center)==1}
}
\value{
a 'numeric' vector. x normalised to center.
}
\description{
Normalise a sample
}
