% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{parse_EL155}
\alias{parse_EL155}
\title{separate vote rows in to columns}
\usage{
parse_EL155(
  votes,
  vote_col = "text",
  start_pos = c(1, 9, 14, 16, 21, 61),
  end_pos = c(7, 12, 14, 19, 59, 131),
  col_names = c("machine", "ballot_style", "marker", "choice_id", "choice_name",
    "contest_name")
)
}
\arguments{
\item{votes}{data frame with row}

\item{vote_col}{column name (in characters) where the votes are}

\item{start_pos}{numerical vector of start char positions}

\item{end_pos}{numerical vector of end char positions}

\item{col_names}{names of variables for each column}
}
\description{
Use fwf to separate out colums in vote rows. Entries need to be votes, thus
they need to be a product of get_precinct_range()
Data looks like
\code{5123906    2 *   10 Nikki R Haley                           GOVERNOR}
}
\examples{
data(G2016_Dillon_EL155)
dillon_p <- get_precinct(G2016_Dillon_EL155)
wprecinct <- add_precinct(G2016_Dillon_EL155, dillon_p)
df <- parse_EL155(wprecinct)
}
