% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballotDef.R
\name{plot.ballotSortDef}
\alias{plot.ballotSortDef}
\title{plot.ballotSortDef}
\usage{
\method{plot}{ballotSortDef}(x, ...)
}
\arguments{
\item{x}{a 'ballotSortDef' object}

\item{...}{extra plot arguments}
}
\description{
plot.ballotSortDef
}
