% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.densityJobDef}
\alias{plot.densityJobDef}
\title{plot.densityJobDef}
\usage{
\method{plot}{densityJobDef}(x, ...)
}
\arguments{
\item{x}{a 'densityJobDef' object}

\item{...}{extra plot arguments}
}
\description{
plot.densityJobDef
}
