% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-.nb.R
\name{poly2nb.funk}
\alias{poly2nb.funk}
\title{poly2nb.funk}
\usage{
poly2nb.funk(sp, LINKFUN = queenLinkFun)
}
\arguments{
\item{sp}{a 'SpatialPolygonsDataFrame' object}

\item{LINKFUN}{a 'function' to determine link status.}
}
\description{
poly2nb.funk
}
