% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{read_EL155}
\alias{read_EL155}
\title{Read a EL155 file}
\usage{
read_EL155(path = "build/input/SC_2010Gen/Allendale/EL155",
  cname = "Allendale", precinct_regex = "(RUN DATE|ELECTION ID:)")
}
\arguments{
\item{path}{path to raw file}

\item{cname}{county name}

\item{precinct_regex}{regex that identifies precinct (a data row)}
}
\description{
Read a EL155 file
}
\examples{
allendale <- read_EL155("build/input/SC_2010Gen/Allendale/EL155", "Allendale")
}
