% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{read_format_EL155}
\alias{read_format_EL155}
\title{Wrapper from path to formatted data}
\usage{
read_format_EL155(paths)
}
\arguments{
\item{paths}{a vector of paths of the county-directory}
}
\description{
Use the step-by-step parsing  functions to generate all votes in a race
}
