% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionTag.R
\name{regionTag}
\alias{regionTag}
\title{regionTag}
\usage{
regionTag(x, rTitle = NULL, rLayerTag = NULL, rMask = NULL)
}
\arguments{
\item{x}{a 'regionTag' or 'character'}

\item{rTitle}{a 'character' title}

\item{rLayerTag}{a 'layerTag' key}

\item{rMask}{a 'vector' defining the subset of the layerTag layer that forms the region.}
}
\description{
regionTag
}
