% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballotTag.R
\name{reportAndCollate.ballotTag}
\alias{reportAndCollate.ballotTag}
\title{reportAndCollate.ballotTag}
\usage{
\method{reportAndCollate}{ballotTag}(x, SBREPORTFUN, COLLATEFUN, ...)
}
\description{
reportAndCollate.ballotTag
}
