% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAndCollate.R
\name{reportAndCollate}
\alias{reportAndCollate}
\title{reportAndCollate}
\usage{
reportAndCollate(sb, SBREPORTFUN, COLLATEFUN = c, ...)
}
\arguments{
\item{sb}{a 'subballot' object}

\item{SBREPORTFUN}{a 'function' that takes a sub-ballot matrix and returns a report}

\item{COLLATEFUN}{a 'function' that takes ... and returns collated results}

\item{...}{extra arguments for SBREPORTFUN}
}
\description{
reportAndCollate
}
