% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sample_nvoters}
\alias{sample_nvoters}
\alias{sample_fvoters}
\title{Take a small sample of voters}
\usage{
sample_nvoters(tbl, n)

sample_fvoters(tbl, frac)
}
\arguments{
\item{tbl}{the dataset that has the columns \code{elec} and \code{voter_id}
(useful if a long version)}

\item{n}{Number of voters to sample}

\item{frac}{Fraction of voters to sample. Alternative to \code{n}}
}
\description{
Custom wrapper around \code{dplyr::sample_n} and \code{dplyr::sample_frac}.
Instead of taking samples of rows, it takes a sample of voters defined by
\code{elec, voter_id}. Useful for long-form datasets where each row is an
vote for a particular office by a person.
}
