% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{sample_to_color}
\alias{sample_to_color}
\title{sample_to_color}
\usage{
sample_to_color(sample, center = mean(sample))
}
\arguments{
\item{sample}{a 'numeric' vector with 1 element per electoral area.}

\item{n}{number of colour bands}
}
\description{
sample_to_color
}
