% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbBorghesiPlot}
\alias{sbBorghesiPlot}
\title{sbBorgesiPlot}
\usage{
sbBorghesiPlot(sb, norm = F, ...)
}
\arguments{
\item{sb}{A sub-ballot}

\item{norm}{'logical' value. Default FALSE. Whether sample is to be normalised.}

\item{...}{extra arguments for plot function}
}
\description{
sbBorgesiPlot
}
