% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbChisqTest}
\alias{sbChisqTest}
\title{sbChisqTest}
\usage{
sbChisqTest(sb, ...)
}
\arguments{
\item{sb}{A sub-ballot}

\item{...}{extra arguments passed on to density function.}
}
\value{
chisq test result of sample of sub-ballot
}
\description{
sbChisqTest
}
