% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbDensity}
\alias{sbDensity}
\title{sbDensity}
\usage{
sbDensity(sb, ...)
}
\arguments{
\item{sb}{A sub-ballot}

\item{...}{extra arguments passed on to density function.}
}
\value{
x location of density peak of sample of sub-ballot
}
\description{
sbDensity
}
