% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbDensityGen}
\alias{sbDensityGen}
\title{sbDensityGen}
\usage{
sbDensityGen(norm = TRUE, ...)
}
\arguments{
\item{norm}{'logical' value. Default FALSE. Whether sample is to be normalised.}

\item{...}{extra arguments for density()}
}
\description{
sbDensityGen
}
