% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbDensityPeakY}
\alias{sbDensityPeakY}
\title{sbDensityPeakY}
\usage{
sbDensityPeakY(sb, ...)
}
\arguments{
\item{sb}{A sub-ballot}

\item{...}{extra arguments passed on to density function.}
}
\value{
y location of density peak of sample of sub-ballot
}
\description{
sbDensityPeakY
}
