% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbKsTest}
\alias{sbKsTest}
\title{sbKsTest}
\usage{
sbKsTest(sb, y = pnorm, ...)
}
\arguments{
\item{sb}{A sub-ballot}

\item{y}{the distribution or measurement to test against. Default is pnorm()}

\item{...}{extra parameters for ks.test()}
}
\description{
sbKsTest
}
