% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbPopMean}
\alias{sbPopMean}
\title{sbPopMean
Calculate the population mean of a sub-ballot}
\usage{
sbPopMean(sb)
}
\arguments{
\item{sb}{A sub-ballot}
}
\value{
The population mean
}
\description{
sbPopMean
Calculate the population mean of a sub-ballot
}
