% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbPopSd}
\alias{sbPopSd}
\title{sbPopSd
Calculate the population standard deviation of a sub-ballot}
\usage{
sbPopSd(sb)
}
\arguments{
\item{sb}{A sub-ballot}
}
\value{
The population standard deviation
}
\description{
sbPopSd
Calculate the population standard deviation of a sub-ballot
}
