% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbReportGen}
\alias{sbReportGen}
\title{sbReportGen}
\usage{
sbReportGen(
  SBREPORTFUNS = list(chisqTest = sbChisqTest, popMean = sbPopMean, popSd = sbPopSd,
    densityPeakX = sbDensityPeakX, densityPeakY = sbDensityPeakY, skewness = sbSkewness,
    kurtosis = sbKurtosis)
)
}
\arguments{
\item{SBREPORTFUNS}{a list of functions}
}
\description{
sbReportGen
}
