% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbSampMean}
\alias{sbSampMean}
\title{sbSampMean
Calculate the sample mean of a sub-ballot}
\usage{
sbSampMean(sb)
}
\arguments{
\item{sb}{A sub-ballot}
}
\value{
The sample mean
}
\description{
sbSampMean
Calculate the sample mean of a sub-ballot
}
