% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subballot.R
\name{sbSimpleTest}
\alias{sbSimpleTest}
\title{sbSimpleTest
A simple test developed for fast computation.}
\usage{
sbSimpleTest(sb, ...)
}
\arguments{
\item{sb}{A sub-ballot}

\item{...}{extra arguments passed on to density function.}
}
\value{
simple test result of sample of sub-ballot
}
\description{
sbSimpleTest
A simple test developed for fast computation.
}
