% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_counties}
\alias{sc_counties}
\title{Information about South Carolina counties}
\format{\describe{
  \item{county}{name of the county, in sentence case (Aiken, Anderson)}
  \item{sol_race_code}{weight of the diamond, in carats}
  ...
}}
\usage{
sc_counties
}
\description{
Some district information at the county level
}
\keyword{datasets}
