% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballotTag.R
\name{set.bRegionTag}
\alias{set.bRegionTag}
\title{set.bRegionTag}
\usage{
set.bRegionTag(x, bRegionTag)
}
\arguments{
\item{x}{a 'ballotTag'}

\item{bRegionTag}{a 'regionTag'}
}
\description{
set.bRegionTag
}
