% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerTag.R
\name{set.linkSearchCol}
\alias{set.linkSearchCol}
\title{set.linkSearchCol}
\usage{
set.linkSearchCol(x, linkSearchCol)
}
\arguments{
\item{x}{a 'layerTag'}

\item{linkSearchCol}{a reference to the field of the shapefile referenced by #'@export}
}
\description{
set.linkSearchCol
}
