% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerTag.R
\name{set.refInfo}
\alias{set.refInfo}
\title{set.refInfo}
\usage{
set.refInfo(x, refInfo)
}
\arguments{
\item{x}{a 'layerTag'}

\item{refInfo}{a 'character' roxygen2 @reference section}
}
\description{
set.refInfo
}
