% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionTag.R
\name{set.rLayerTag}
\alias{set.rLayerTag}
\title{set.rLayerTag}
\usage{
set.rLayerTag(x, rLayerTag)
}
\arguments{
\item{x}{a 'regionTag' tag}

\item{rLayerTag}{a 'layerTag' key}
}
\description{
set.rLayerTag
}
