% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionTag.R
\name{set.rMask}
\alias{set.rMask}
\title{set.rMask}
\usage{
set.rMask(x, rMask)
}
\arguments{
\item{x}{a 'regionTag' tag}

\item{rMask}{a 'vector' defining the subset of the layerTag layer that forms the region.}
}
\description{
set.rMask
}
