% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layerTag.R
\name{set.textLinks}
\alias{set.textLinks}
\title{set.textLinks}
\usage{
set.textLinks(x, textLinks)
}
\arguments{
\item{x}{a 'layerTag'}

\item{textLinks}{a 'character' vector defining links (ferries / bridges / etc)}
}
\description{
set.textLinks
}
