% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{slim_bind}
\alias{slim_bind}
\title{Bind a list of dataframes, but with only a subset of variables}
\usage{
slim_bind(df_list, regex, default = c("elec", "county", "precinct_id",
  "ballot_style", "voter_id"))
}
\arguments{
\item{df_list}{A list of dataframes to bind.}

\item{regex}{A regular expression to}

\item{default}{character vector of variables (like identifiers to always keep)}
}
\description{
Wrapper to map_dfr but specifies variables to use. This is useful when the underlying data
is large and has lots of variables that are not necessary.
}
