% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{spatialPlot}
\alias{spatialPlot}
\title{spatialPlot}
\usage{
spatialPlot(x, sample = NULL, ...)
}
\arguments{
\item{x}{a 'character' ballotTag}

\item{sample}{a 'numeric' vector with 1 element per electoral area.}

\item{...}{extra plot parameters}
}
\description{
spatialPlot
}
