% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{std_race}
\alias{std_race}
\title{Standardize race variables to proper names with leading 7-character codes}
\usage{
std_race(vec)
}
\arguments{
\item{vec}{Vector of unstandardized race names}
}
\description{
Standardize race variables to proper names with leading 7-character codes
}
\examples{
vec <- c("CON0001 House 1", "CONG007 House 7", "U. S. Senator")
std_race(vec)

}
