% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{std_ref_option}
\alias{std_ref_option}
\title{Standardize options in referendum race}
\usage{
std_ref_option(vec)
}
\arguments{
\item{vec}{Vector of candidate / vote options}
}
\description{
Standardize options in referendum race
}
