% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{std_sc_first}
\alias{std_sc_first}
\title{Standardize race with first letter after code}
\usage{
std_sc_first(vec, use, code_regex, code_replace)
}
\arguments{
\item{vec}{a character vector that contains some values to be used in code}

\item{use}{a logical vector of the same length as \code{vec} that is
\code{TRUE} when the index is to be transformed, and \code{FALSE} otherwise

raw <- c("CSB West Ashley", "CSB East Cooper", "County Council")
use <- c(TRUE, TRUE, FALSE)
std_sc_first(raw, use, "CSB ", "SCH")}
}
\description{
Standardize race with first letter after code
}
