% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{std_sc_solicit}
\alias{std_sc_solicit}
\title{Standardize solicitor contest with district number, based on county}
\usage{
std_sc_solicit(vec, is_solicit, county, data = sc_counties)
}
\arguments{
\item{vec}{a character vector that contains some solicitor district values}

\item{county}{A vector of strings, south carolina counties.}

\item{data}{A data that matches counties to solicitor district keys. Provided in package.}

\item{is_solicitor}{a logical vector of the same length as \code{vec} that is
\code{TRUE} when the index is a solicitor, and \code{FALSE} otherwise}
}
\description{
Standardize solicitor contest with district number, based on county
}
\examples{
raw <- c("Solicitor Circuit District 14", "Solicitor 10th Circuit", "County Council")
is_solicit <- c(TRUE, TRUE, FALSE)
county <- c("Anderson", "Jasper", "Charleston")
std_sc_solicit(raw, is_solicit, county)

}
