% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{std_yes_no}
\alias{std_yes_no}
\title{Standardize options in referendum race}
\usage{
std_yes_no(vec)
}
\arguments{
\item{vec}{Vector of candidate / vote options}
}
\description{
Standardize options in referendum race
}
\examples{
std_yes_no(c("No", "Opposed to Question", "Opposed", "W/I ANY OPPOSED"))

}
