% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{stdnum_yesno}
\alias{stdnum_yesno}
\title{Standardize Yes / No / NA to -1, 1, 0.}
\usage{
stdnum_yesno(vec)
}
\arguments{
\item{vec}{long values of Yes/Nos Must have been passed through filter_existing,
otherwise all NAs are coded as 0s}
}
\description{
Standardize Yes / No / NA to -1, 1, 0.
}
