% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_candidates.R
\name{stwide_to_numeric}
\alias{stwide_to_numeric}
\title{Encode party of statewide offices in South Carolina}
\usage{
stwide_to_numeric(tbl, year)
}
\arguments{
\item{tbl}{wide table with columns like \env{GOV0000}.}

\item{year}{year of general election}
}
\description{
Recode each year's statewide elections to numeric by hand. Because we know these were on the ballot for everyone, numeric codes can be determined as a 0
}
