% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xLoad.R
\name{xLoad}
\alias{xLoad}
\title{xLoad}
\usage{
xLoad(x, envir = environment())
}
\arguments{
\item{x}{a 'character' describing a file to attempt loading.}

\item{envir}{an 'environment' in which to load the specified file.}
}
\description{
Load on exist.
}
