% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{add_precinct}
\alias{add_precinct}
\title{Add precinct to each row of vote dataset, and remove other headers}
\usage{
add_precinct(votes, pkey)
}
\arguments{
\item{votes}{EL155 dataset of votes, product of read_EL155()}

\item{pkey}{precinct key, product of get_precinct_range()}
}
\description{
Use the non-equi join in data.table to identify precinct. row i is in
precinct j if i's ID is in between p_start_id and p_end_id of precint j.
}
\examples{
allendale <- read_EL155("build/input/SC_2010Gen/Allendale/EL155", "Allendale")
ald_p <- get_precinct_range(allendale)
wprecinct <- add_precinct(allendale, ald_p)
}
