% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_to_party.R
\name{join_1col}
\alias{join_1col}
\title{Join with candidate data with single district identifier}
\usage{
join_1col(tbl, cand, vote_col, join_cols_tbl, join_cols_cand)
}
\arguments{
\item{tbl}{A wide dataset with the contest}

\item{join_col}{NSE code to indicate variable that is being joined (to ballot_name in cands)}

\item{type}{NSE code for the contest. Will also be used to filter \env{cand} if not null.}

\item{...}{geographic variables to segment on

#' This assumes that all rows inlcuded in \code{tbl} did have that election to choose and
that the candidate table is complete; therefore NAs are coded to 0 (abstain).}
}
\description{
For example, used for referenda where there is only one "seat"
}
