% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_coding.R
\name{join_dist}
\alias{join_dist}
\alias{join_county}
\alias{join_countydist}
\title{Join by a single district identifier (congerss, state hosue) or county}
\usage{
join_dist(tbl, cands, office)

join_county(tbl, cands, office)

join_countydist(tbl, cands, office)
}
\arguments{
\item{tbl}{variables must be of the type '{office}_{vote}', '{office}_{dist}' (if dist) and
county
for vote name and district, respectively.}

\item{cands}{A table of candidates. Must have variables \code{elec}, \code{dist},
\code{county}, \code{ballot_name}, \code{party_num}, \code{contest_type}}

\item{office}{NSE short form of the office in question}
}
\description{
Calls \link{join_1col} internally
}
