% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{list_to_df}
\alias{list_to_df}
\title{Change list of row-wise vectors to df}
\usage{
list_to_df(lst, votes)
}
\arguments{
\item{lst}{list, each item is a row}

\item{votes}{data frame of votes}
}
\description{
Note: time consuming
}
