% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert parsed list to data frame}
\usage{
list_to_df(lst, votes)
}
\arguments{
\item{lst}{list which is a product of parse_EL155}

\item{votes}{dataset of votes to be combined to parsed list. Must be in right ordersame order}
}
\description{
Convert parsed list to data frame
}
\examples{
allendale <- read_EL155("build/input/SC_2010Gen/Allendale/EL155", "Allendale")
ald_p <- get_precinct_range(allendale)
wprecinct <- add_precinct(allendale, ald_p)
lst <- parse_EL155(wprecinct)
df <- list_to_df(lst, wprecinct)
}
