% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SC.R
\name{read_EL155}
\alias{read_EL155}
\title{Read a EL155 file}
\usage{
read_EL155(path = "build/input/SC_2010/Allendale/EL155",
  cname = "Allendale")
}
\arguments{
\item{path}{path to raw file}

\item{cname}{county name}
}
\description{
Read a EL155 file
}
