% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_candidates.R
\name{recode_abbrv}
\alias{recode_abbrv}
\title{Translate abbreviation codes to labels}
\usage{
recode_abbrv(vec, linebreak = FALSE)
}
\arguments{
\item{vec}{a character vector of variable labels}

\item{linebreak}{whether or not to linebreak labels}
}
\description{
Translate abbreviation codes to labels
}
