% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAndCollate.R
\name{reportAndCollate}
\alias{reportAndCollate}
\title{reportAndCollate}
\usage{
reportAndCollate(x, key = "!V", SBREPORTFUN = sbSimpleTest, COLLATEFUN = c)
}
\arguments{
\item{SBREPORTFUN}{a 'function' that takes a sub-ballot matrix and returns a report}

\item{COLLATEFUN}{a 'function' that collates results}

\item{sb}{a 'subballot' object}
}
\description{
reportAndCollate
}
