% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{std_contest}
\alias{std_contest}
\title{Standardize contest variables}
\usage{
std_contest(vec, .type = NULL)
}
\arguments{
\item{vec}{Vector of unstandardized contest names}
}
\description{
Change a vector of contest names to o proper names with leading 7-character codes
}
\examples{
vec <- c(
  "CON0001 House 1", "CONG007 House 7", "CONGR02 U.S. House of Rep Dist 2",
  "CNG0003 U S House of Representatives Dis",
  "CONO006 U S House of Rep Dist 6	",
  "U S House of Rep Dist 6",
  "CON02 U S House of Representatives Distr",
  "CON0001 House 1", "CONG007 House 7",
  "U. S. Senator",
  "President", "PREsident",
  "Straight Party",
  "U. S. Senator", "CCNL001 Council 1",
  "CCNL001 Council 1", "CCD0001 Council 1"
)
std_contest(vec)
}
