% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{std_sc_solicit}
\alias{std_sc_solicit}
\title{Standardize race with district number}
\usage{
std_sc_solicit(vec, is_solicit)
}
\arguments{
\item{vec}{a character vector that contains some solicitor district values}

\item{is_solicitor}{a logical vector of the same length as \code{vec} that is
\code{TRUE} when the index is a solicitor, and \code{FALSE} otherwise

raw <- c("Solicitor Circuit District 14", "Solicitor 10th Circuit", "County Council")
is_solicit <- c(TRUE, TRUE, FALSE)
std_sc_solicit(raw, is_solicit)}
}
\description{
Standardize race with district number
}
